requirements.txt
python-dotenv==1.1.0
setuptools==78.1.0
wheel==0.45.1
